added ending, tidied up
added functionality in parts.lua, where a random number between  450 and 500 inclusive is decided as the ending measure number.
added graphics methods in order to reformat redraw() to adhere to style. added resetSong() but i am unsure if resetSong() is needed, as i have assumed the user will retrigger the song using k3, and then k2.